third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFEConvolveMatrixElement-svgdom-bias-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'bias' property of the SVGFEConvolveMatrixElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS convolveMatrixElement.bias.baseVal is 0.5
PASS convolveMatrixElement.bias.baseVal is 0
PASS successfullyParsed is true

TEST COMPLETE
